[AD] Special-case struct_extract in activity analysis
Previously, extracting a non-differentiable (and @noDerivative) stored 
property from a differentiable struct caused activity analysis to 
propagate VARIED activity from the parent struct into the extracted 
value. This then causes a crash when attempting to look up the cotangent 
vector of the nondifferentiable type.

I consider this change to be a workaround for a more fundamental design 
issue. Figuring out what to differentiate should be an inside-out 
process (one that starts with the differentiability properties of the 
innermost function calls and ends up with a check that the derivatives 
the user requested actually exist) not an outside-in one that relies on 
activity analysis to correctly propagate nondifferentiability. That is, 
activity analysis should be purely a performance optimization over and 
above what the Swift and LLVM optimizers are able to do, and the 
semantics of what to differentiate should be driven by the 
differentiability of called functions.